Fix Computer.MemorySize deserialization for large values
The Jamf Protect API returns memorySize in scientific notation
(e.g. 8.589934592E9 for 8GB). Go's json.Unmarshal cannot decode
scientific notation floats into int64, causing GetComputer to fail
with "cannot unmarshal number" errors.

Change MemorySize from *int64 to *float64 to handle this correctly.

+++ jamfprotect/computer.go
@@ -107,7 +107,7 @@ type Computer struct {
 	OSPatch                 *int64        `json:"osPatch"`
 	Arch                    *string       `json:"arch"`
 	CertID                  *string       `json:"certid"`
-	MemorySize              *int64        `json:"memorySize"`
+	MemorySize              *float64      `json:"memorySize"`
 	OSString                *string       `json:"osString"`
 	KernelVersion           *string       `json:"kernelVersion"`
 	InstallType             *string       `json:"installType"`
